requirements.txt
google-adk>=0.1.0
boto3>=1.34.0
databricks-sdk>=0.18.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.0
aiofiles>=23.2.0
click>=8.1.0